chore: share the native binary verifier between UUAV and RustSegment
Per review: the verifier, Gate B and their self-tests move to
scripts/native-lock/ and take the plugin's lock with --lock. Each lock
carries top-level name/docs for output prefixes and README pointers; the
FFmpeg checks run only for targets that ship FFmpeg artifacts. The MSVC
probe, recursive completeness check and Gate B outcome file now apply to
UUAV too, whose publish summary gains the per-target Gate B table. Only
build-canonical.sh and repro-gate.sh stay per plugin; build-canonical.sh
now explains what "canonical" means.
