+++ .github/workflows/rust-segment-native.yml
@@ -87,10 +87,10 @@ jobs:
         run: bash scripts/rust-segment/repro-gate.sh macos-universal
 
       - name: Gate B - reproduce the committed binaries, toolchain permitting
-        run: python3 scripts/rust-segment/reproduces-lock.py macos-universal --toolchain toolchain-macos.txt
+        run: python3 scripts/native-lock/reproduces-lock.py --lock "$LOCK" macos-universal --toolchain toolchain-macos.txt
 
       - name: Compare the fresh build against the lock
-        run: python3 scripts/rust-segment/verify-binaries.py --report macos-universal
+        run: python3 scripts/native-lock/verify-binaries.py --lock "$LOCK" --report macos-universal
 
       # always(): a failed gate is exactly the round whose hashes and toolchain need reading.
       - uses: actions/upload-artifact@v4
@@ -139,7 +139,7 @@ jobs:
             exit 1
           fi
 
-      # rustc locates link.exe itself, so no vcvars shell names the toolset; vswhere does, as verify-binaries.py's msvc probe reads it back.
+      # rustc locates link.exe itself, so no vcvars shell names the toolset; vswhere does, as the shared verifier's msvc probe reads it back.
       - name: Record the toolchain actually used
         run: |
           set -euo pipefail
@@ -158,10 +158,10 @@ jobs:
         run: bash scripts/rust-segment/repro-gate.sh windows-x86_64
 
       - name: Gate B - reproduce the committed binaries, toolchain permitting
-        run: python scripts/rust-segment/reproduces-lock.py windows-x86_64 --toolchain toolchain-windows.txt
+        run: python scripts/native-lock/reproduces-lock.py --lock "$LOCK" windows-x86_64 --toolchain toolchain-windows.txt
 
       - name: Compare the fresh build against the lock
-        run: python scripts/rust-segment/verify-binaries.py --report windows-x86_64
+        run: python scripts/native-lock/verify-binaries.py --lock "$LOCK" --report windows-x86_64
 
       # always(): see the macOS job.
       - uses: actions/upload-artifact@v4
@@ -216,7 +216,7 @@ jobs:
             echo '```'
             echo
             echo '> [!WARNING]'
-            echo "> **Nothing was committed** — this workflow has \`contents: read\` and only verifies. If this build is meant to ship, follow the relock steps in \`Explorer/Assets/Plugins/RustSegment/README.md\`: commit the binaries from this run's artifacts, then run \`scripts/rust-segment/verify-binaries.py --update --only <target> --toolchain toolchain-<os>.txt\`."
+            echo "> **Nothing was committed** — this workflow has \`contents: read\` and only verifies. If this build is meant to ship, follow the relock steps in \`Explorer/Assets/Plugins/RustSegment/README.md\`: commit the binaries from this run's artifacts, then run \`scripts/native-lock/verify-binaries.py --lock $LOCK --update --only <target> --toolchain toolchain-<os>.txt\`."
           } >> "$GITHUB_STEP_SUMMARY"
 
       - uses: actions/upload-artifact@v4
