+++ docs/build-and-ci.md
@@ -23,7 +23,7 @@ Two Rust plugins ship prebuilt native binaries committed to Git LFS: UUAV (`Expl
 | UUAV | `scripts/uuav/uuav-binaries.lock.json` | `uuav-verify.yml` | `uuav-native.yml` (`build-uuav-native`) |
 | RustSegment | `scripts/rust-segment/rust-segment-binaries.lock.json` | `rust-segment-verify.yml` | `rust-segment-native.yml` (`build-rust-segment-native`) |
 
-The verify workflow re-hashes the shipped binaries and their build inputs against the lock; the rebuild workflow builds on hosted runners under the pinned toolchain and fails unless two builds produce identical bytes, then hands the binaries back as artifacts for a human to commit and relock. Neither commits. The design and the relock steps are documented in each plugin's README ([UUAV](../Explorer/Assets/Plugins/UUAV/README.md#ci-verification-of-the-shipped-binaries), [RustSegment](../Explorer/Assets/Plugins/RustSegment/README.md#ci-verification-of-the-shipped-binaries)); [`docs/uuav.md`](uuav.md) explains why the binaries are committed rather than built on demand.
+The verifier, Gate B and their self-tests live once under `scripts/native-lock/` and take the plugin's lock with `--lock`; each plugin keeps only its build and Gate A shell scripts. The verify workflow re-hashes the shipped binaries and their build inputs against the lock; the rebuild workflow builds on hosted runners under the pinned toolchain and fails unless two builds produce identical bytes, then hands the binaries back as artifacts for a human to commit and relock. Neither commits. The design and the relock steps are documented in each plugin's README ([UUAV](../Explorer/Assets/Plugins/UUAV/README.md#ci-verification-of-the-shipped-binaries), [RustSegment](../Explorer/Assets/Plugins/RustSegment/README.md#ci-verification-of-the-shipped-binaries)); [`docs/uuav.md`](uuav.md) explains why the binaries are committed rather than built on demand.
 
 ## GitHub Workflow
